vps-ocr-service/requirements.txt
# FastAPI and ASGI server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# OCR and Document Processing
paddleocr==2.7.3
paddlepaddle==2.6.1  # CPU version for 2vCPU VPS
pdf2image==1.17.0
Pillow==10.2.0

# Embeddings and ML
sentence-transformers==2.3.1
torch==2.2.0+cpu  # CPU-only PyTorch (lighter)
--extra-index-url https://download.pytorch.org/whl/cpu

# Database
psycopg2-binary==2.9.9
asyncpg==0.29.0  # For async PostgreSQL

# HTTP client
aiohttp==3.9.1
httpx==0.26.0

# Utilities
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Logging
loguru==0.7.2
